requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
altair==4.2.2
argcomplete==1.10.3
async-timeout==4.0.2
attrs==23.1.0
beautifulsoup4==4.8.2
blinker==1.6.2
cachetools==5.3.0
certifi==2022.12.7
chardet==3.0.4
charset-normalizer==3.1.0
click==8.1.3
compressed-rtf==1.0.6
decorator==5.1.1
docx2txt==0.8
ebcdic==1.1.1
entrypoints==0.4
extract-msg==0.28.7
frozenlist==1.3.3
gitdb==4.0.10
GitPython==3.1.31
idna==3.4
IMAPClient==2.1.0
importlib-metadata==6.6.0
Jinja2==3.1.2
joblib==1.2.0
jsonschema==4.17.3
lxml==4.9.2
markdown-it-py==2.2.0
MarkupSafe==2.1.2
mdurl==0.1.2
multidict==6.0.4
nltk==3.8.1
numpy==1.24.3
olefile==0.46
openai==0.27.4
packaging==23.1
pandas==1.5.3
pdfminer.six==20191110
Pillow==9.5.0
protobuf==3.20.3
pyarrow==11.0.0
pycryptodome==3.17
pydeck==0.8.1b0
Pygments==2.15.1
Pympler==1.0.1
pyrsistent==0.19.3
python-dateutil==2.8.2
python-dotenv==1.0.0
python-pptx==0.6.21
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
regex==2023.3.23
requests==2.28.2
rich==13.3.4
six==1.12.0
smmap==5.0.0
sortedcontainers==2.4.0
soupsieve==2.4.1
SpeechRecognition==3.8.1
streamlit==1.21.0
streamlit-chat==0.0.2.2
textract==1.6.5
tiktoken==0.3.3
toml==0.10.2
toolz==0.12.0
tornado==6.3.1
tqdm==4.65.0
typing_extensions==4.5.0
tzdata==2023.3
tzlocal==4.3
urllib3==1.26.15
validators==0.20.0
watchdog==3.0.0
xlrd==1.2.0
XlsxWriter==3.1.0
yarl==1.9.1
zipp==3.15.0
